+++ src/polynx/frame.py
@@ -13,8 +13,7 @@ class DataFrame:
             _pl = super().__getattribute__("_pl")
             attr = getattr(_pl, name)
             if callable(attr):
-                def wrapped(*args, **kwargs):
-                    from .wrapper import wrap, unwrap
+                def wrapped(*args, **kwargs):                    
                     return wrap(attr(*[unwrap(a) for a in args], **{k: unwrap(v) for k, v in kwargs.items()}))
                 return wrapped
             return attr  
@@ -28,3 +27,6 @@ class DataFrame:
 
     def to_polars(self):
         return self._pl
+    
+    def to_pandas(self):
+        return self._pl.to_pandas()